+++ src/main/java/sarasa/wantedinternship/dto/request/SignUpDto.java
@@ -1,4 +1,4 @@
-package sarasa.wantedinternship.dto;
+package sarasa.wantedinternship.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Size;
