+++ src/main/java/sarasa/wantedinternship/mapper/MemberMapper.java
@@ -2,7 +2,7 @@ package sarasa.wantedinternship.mapper;
 
 import org.mapstruct.Mapper;
 import sarasa.wantedinternship.domain.entity.Member;
-import sarasa.wantedinternship.dto.SignUpDto;
+import sarasa.wantedinternship.dto.request.SignUpDto;
 
 @Mapper(componentModel = "spring")
 public interface MemberMapper {
